Improvements on api-utils
